+++ src/Convert/Public/ConvertTo-Clixml2.ps1
@@ -13,7 +13,7 @@
 
     .EXAMPLE
         $string = 'A string'
-        ConvertTo-Clixml -InputObject $string
+        ConvertTo-Clixml2 -InputObject $string
 
 <Objs Version="1.1.0.1" xmlns="http://schemas.microsoft.com/powershell/2004/04">
   <S>A string</S>
@@ -21,7 +21,7 @@
 
     .EXAMPLE
         $string = 'A string'
-        $string | ConvertTo-Clixml
+        $string | ConvertTo-Clixml2
 
 <Objs Version="1.1.0.1" xmlns="http://schemas.microsoft.com/powershell/2004/04">
   <S>A string</S>
@@ -30,7 +30,7 @@
     .EXAMPLE
         $string1 = 'A string'
         $string2 = 'Another string'
-        ConvertTo-Clixml -InputObject $string1,$string2
+        ConvertTo-Clixml2 -InputObject $string1,$string2
 
 <Objs Version="1.1.0.1" xmlns="http://schemas.microsoft.com/powershell/2004/04">
   <S>A string</S>
@@ -42,7 +42,7 @@
     .EXAMPLE
         $string1 = 'A string'
         $string2 = 'Another string'
-        $string1,$string2 | ConvertTo-Clixml
+        $string1,$string2 | ConvertTo-Clixml2
 
 <Objs Version="1.1.0.1" xmlns="http://schemas.microsoft.com/powershell/2004/04">
   <S>A string</S>
@@ -55,10 +55,10 @@
         [String[]]
 
     .LINK
-        http://convert.readthedocs.io/en/latest/functions/ConvertTo-Clixml/
+        http://convert.readthedocs.io/en/latest/functions/ConvertTo-Clixml2/
 #>
-function ConvertTo-Clixml {
-    [CmdletBinding(HelpUri = 'http://convert.readthedocs.io/en/latest/functions/ConvertTo-Clixml/')]
+function ConvertTo-Clixml2 {
+    [CmdletBinding(HelpUri = 'http://convert.readthedocs.io/en/latest/functions/ConvertTo-Clixml2/')]
     param
     (
         [Parameter(